Make text smaller and limit width of index page
Changed max-width of content on index page as the content is not wide itself
and having table names and item counts on opposite sides is hard to align
on wide monitors.

Text size changed to something that allows to see more at the same time while
still being comfortably big to read.